LPS-164683 LPS-164686 portal-search-web: Refactor SearchBarConfigurationSuggestions to separate files
